Working mechanism of Sun Tracking System
Each LDR (Light Dependent Resistor) placed on the board so that it could point to the sun to varying time zone. This allows each LDR to effectively detect varying levels of solar illumination depending on the specific situation. As a result, the different LED light gets illuminated to the situation of the sun.